+++ src/sstat.h
@@ -9,12 +9,12 @@
 
 class MultinomialSeqStat {
 private:
-    int num_char_;
-    int num_taxa_;
-    double test_statistic_;
+    unsigned long num_char_;
+    unsigned long num_taxa_;
+    long double test_statistic_;
     
     std::vector<Sequence> seqs_;
-    std::vector< std::pair <std::string, int> > patterns_and_counts_;
+    std::vector< std::pair <std::string, unsigned long> > patterns_and_counts_;
     
     bool checked_aligned ();
     void collect_site_patters ();
@@ -22,7 +22,7 @@ private:
     
 public:
     explicit MultinomialSeqStat (std::vector<Sequence>& seqs);
-    double get_test_statistic () const;
+    long double get_test_statistic () const;
     
     //~MultinomialSeqStat();
 };
